# src/main/java/com/effective_mobile/card_management/docs/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
  description:  backend-приложение на Java (Spring Boot) для управления банковскими картами
servers:
  - url: http://localhost:8080
    description: Локальный сервер
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
        cardNumberMasked:
          type: string
          description: Замаскированный номер карты
        owner:
          type: string
          description: Имя владельца
        expiryDate:
          type: string
          format: date
          description: Срок действия карты (YYYY-MM-DD)
        status:
          type: string
          description: Статус карты (например, ACTIVE, BLOCKED, EXPIRED)
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
          format: double
          description: Баланс карты
    CardCreateDto:
      type: object
      properties:
        owner:
          type: string
          description: Имя владельца
        expiryDate:
          type: string
          format: date
          description: Срок действия карты (YYYY-MM-DD)
        balance:
          type: number
          format: double
          description: Начальный баланс
        userId:
          type: integer
          format: int64
          description: ID пользователя, которому принадлежит карта
    CardUpdateDto:
      type: object
      properties:
        status:
          type: string
          description: Новый статус карты (например, ACTIVE, BLOCKED, EXPIRED)
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
...
paths:
...
  /admin/cards:
    post:
      security:
        - bearerAuth: []
      summary: Регистрация карты пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreateDto'
      responses:
        '200':
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Неверный запрос (например, неверный формат данных)
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Доступ запрещен
        '500':
          description: Внутренняя ошибка сервера
          content:
            text/plain:
              schema:
                type: string

    get:
      security:
        - bearerAuth: []
      summary: Получение всех карт
      responses:
        '200':
          description: Список карт получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
        '403':
          description: Доступ запрещен
        '500':
          description: Внутренняя ошибка сервера
          content:
            text/plain:
              schema:
                type: string
  /admin/cards/{id}:
    put:
      security:
        - bearerAuth: []
      summary: Обновление карты пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdateDto'
      responses:
        '200':
          description: Карта успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardDto'
        '400':
          description: Неверный запрос (например, неверный формат данных)
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            text/plain:
              schema:
                type: string

    delete:
      security:
        - bearerAuth: []
      summary: Удаление карты пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
      responses:
        '204':
          description: Карта успешно удалена
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            text/plain:
              schema:
                type: string
  /admin/cards/{id}/block:
    post:
      security:
        - bearerAuth: []
      summary: Блокировка карты пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
      responses:
        '200':
          description: Карта успешно заблокирована
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            text/plain:
              schema:
                type: string
  /admin/cards/{id}/activate:
    post:
      security:
        - bearerAuth: []
      summary: Активация карты пользователя
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID карты
      responses:
        '200':
          description: Карта успешно активирована
        '403':
          description: Доступ запрещен
        '404':
          description: Карта не найдена
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Внутренняя ошибка сервера
          content:
            text/plain:
              schema:
                type: string
...